factor filehashing functions out into separate modeul
