+++ isaw/images/package.py
@@ -8,8 +8,8 @@ from arglogger import arglogger
 from io import BytesIO
 import datetime
 import exiftool
+from filehashing import hash_of_file, safe_copy
 from functools import wraps
-import hashlib
 from PIL.ImageCms import getOpenProfile, getProfileName, profileToProfile, ImageCmsProfile
 import json
 import logging
@@ -64,26 +64,6 @@ for k in IMAGETYPES.keys():
 SIZEPREVIEW = 800, 600
 SIZETHUMB = 128, 128
 
-def hash_of_file(filepath):
-    """
-    generate sha1 hash for a file
-    """
-    with open(filepath, 'rb') as f:
-        return hashlib.sha1(f.read()).hexdigest()
-
-def safe_copy(src, dest, tries=2):
-    """
-    verify checksums on file copy
-    """
-    i = 0
-    while i < tries:
-        hash_src = hash_of_file(src)
-        shutil.copy2(src, dest)
-        hash_dest = hash_of_file(dest)
-        if hash_src == hash_dest:
-            return hash_dest
-        i += 1
-    raise IOError("could not verify safe-copy from {source} to {dest}".format(source=src, dest=dest))
 
 class Package:
     """
